requirements.txt
# Do we need altair[all]
altair
babel
beautifulsoup4
boto3
cachetools
colorama
cookiecutter
duckdb
jinja2
jsonschema[format-nongpl]
markdown
pg8000
phonenumbers
pillow
pydantic
pypdf
python-dotenv
pyyaml
qrcode[pil]
referencing
requests
tqdm
vl-convert-python
weasyprint
